requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
--index-url https://pypi.tuna.tsinghua.edu.cn/simple

annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via starlette
argon2-cffi==25.1.0
    # via pwdlib
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
asyncpg==0.31.0
    # via -r requirements.in
bcrypt==5.0.0
    # via pwdlib
cffi==2.0.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
click==8.3.1
    # via uvicorn
colorama==0.4.6
    # via click
cryptography==46.0.3
    # via pyjwt
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via fastapi-users
fastapi==0.127.0
    # via
    #   -r requirements.in
    #   fastapi-users
fastapi-users==15.0.3
    # via -r requirements.in
geoalchemy2==0.18.1
    # via -r requirements.in
greenlet==3.3.0
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   email-validator
makefun==1.16.0
    # via fastapi-users
numpy==2.4.0
    # via shapely
packaging==25.0
    # via geoalchemy2
psycopg2==2.9.11
    # via -r requirements.in
pwdlib[argon2,bcrypt]==0.3.0
    # via fastapi-users
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.41.5
    # via pydantic
pyjwt[crypto]==2.10.1
    # via fastapi-users
python-bcrypt==0.3.2
    # via -r requirements.in
python-dotenv==1.2.1
    # via -r requirements.in
python-multipart==0.0.21
    # via fastapi-users
shapely==2.1.2
    # via -r requirements.in
sqlalchemy[asyncio]==2.0.36
    # via
    #   -r requirements.in
    #   geoalchemy2
    #   sqlalchemy
starlette==0.50.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uvicorn==0.40.0
    # via -r requirements.in
